chore: send messages asynchronously in per peer spawned tasks

+++ state-chain/multisig-p2p-transport/src/lib.rs
@@ -30,6 +30,7 @@ use std::{
 	net::Ipv6Addr,
 	pin::Pin,
 };
+use tokio::sync::mpsc::{error::TrySendError, Sender};
 
 #[cfg(test)]
 use mockall::automock;
@@ -77,6 +78,7 @@ pub struct RpcRequestHandler<MetaData, P2PNetworkService: PeerNetwork> {
 	/// Runs concurrently in the background and manages receiving (from the senders in
 	/// "p2p_message_rpc_subscribers") and then actually sending P2PEvents to the Rpc subscribers
 	notification_rpc_subscription_manager: SubscriptionManager,
+	message_sender_spawner: sc_service::SpawnTaskHandle,
 	state: Arc<RwLock<P2PValidatorNetworkNodeState>>,
 	p2p_network_service: Arc<P2PNetworkService>,
 	_phantom: std::marker::PhantomData<MetaData>,
@@ -88,7 +90,7 @@ struct P2PValidatorNetworkNodeState {
 	/// Store all local rpc subscriber senders
 	p2p_message_rpc_subscribers:
 		HashMap<SubscriptionId, UnboundedSender<(PeerIdTransferable, Vec<u8>)>>,
-	reserved_peers: BTreeMap<PeerId, (u16, Ipv6Addr)>,
+	reserved_peers: BTreeMap<PeerId, (u16, Ipv6Addr, Sender<Vec<u8>>)>,
 }
 
 /// An abstration of the underlying network of peers.
@@ -187,11 +189,52 @@ pub trait P2PValidatorNetworkNodeRpcApi {
 	) -> Result<bool>;
 }
 
+impl<
+		MetaData: jsonrpc_pubsub::PubSubMetadata + Send + Sync + 'static,
+		PN: PeerNetwork + Send + Sync + 'static,
+	> RpcRequestHandler<MetaData, PN>
+{
+	fn update_peer_mapping(
+		&self,
+		state: &mut std::sync::RwLockWriteGuard<P2PValidatorNetworkNodeState>,
+		peer_id: PeerId,
+		port: u16,
+		ip_address: Ipv6Addr,
+	) -> bool {
+		if let Some((existing_port, existing_ip_address, _message_sender)) =
+			state.reserved_peers.get_mut(&peer_id)
+		{
+			if *existing_port != port || *existing_ip_address != ip_address {
+				*existing_port = port;
+				*existing_ip_address = ip_address;
+				// TODO Check that removing then adding a peer is enough
+				self.p2p_network_service.remove_reserved_peer(peer_id);
+				self.p2p_network_service.reserve_peer(peer_id, port, ip_address);
+				true
+			} else {
+				false
+			}
+		} else {
+			let (sender, mut receiver) = tokio::sync::mpsc::channel(16);
+			state.reserved_peers.insert(peer_id.clone(), (port, ip_address, sender));
+			let p2p_network_service = self.p2p_network_service.clone();
+			self.message_sender_spawner.spawn("cf-peer-message-sender", async move {
+				while let Some(message) = receiver.recv().await {
+					p2p_network_service.write_notification(peer_id, message);
+				}
+			});
+			self.p2p_network_service.reserve_peer(peer_id, port, ip_address);
+			true
+		}
+	}
+}
+
 pub fn new_p2p_validator_network_node<
 	MetaData: jsonrpc_pubsub::PubSubMetadata + Send + Sync + 'static,
 	PN: PeerNetwork + Send + Sync + 'static,
 >(
 	p2p_network_service: Arc<PN>,
+	message_sender_spawner: sc_service::SpawnTaskHandle,
 	subscription_task_executor: impl Spawn + Send + Sync + 'static,
 ) -> (RpcRequestHandler<MetaData, PN>, impl futures::Future<Output = ()>) {
 	let state = Arc::new(RwLock::new(P2PValidatorNetworkNodeState::default()));
@@ -211,7 +254,7 @@ pub fn new_p2p_validator_network_node<
 					&self,
 					peers: Vec<(PeerIdTransferable, u16, Ipv6Addr)>,
 				) -> Result<u64> {
-					let mut peers = peers
+					let peers = peers
 						.into_iter()
 						.map(|(peer_id, port, address)| {
 							Result::Ok((peer_id.try_into()?, (port, address)))
@@ -219,25 +262,19 @@ pub fn new_p2p_validator_network_node<
 						.collect::<std::result::Result<BTreeMap<_, _>, _>>()?;
 
 					let mut state = self.state.write().unwrap();
-					std::mem::swap(&mut state.reserved_peers, &mut peers);
+					state.reserved_peers.retain(|peer_id, _| {
+						if peers.contains_key(peer_id) {
+							true
+						} else {
+							self.p2p_network_service.remove_reserved_peer(*peer_id);
+							false
+						}
+					});
 
 					// TODO: Investigate why adding/removing multiple reserved peers in a single
 					// reserve_peers call doesn't work
-
-					// TODO Check that removing then adding a peer is enough
-
-					for (peer_id, _) in peers.iter().filter(|(peer_id, port_addr)| {
-						state.reserved_peers.get(peer_id) != Some(port_addr)
-					}) {
-						self.p2p_network_service.remove_reserved_peer(*peer_id);
-					}
-
-					for (peer_id, (port, addr)) in state
-						.reserved_peers
-						.iter()
-						.filter(|(peer_id, addr_port)| peers.get(peer_id) != Some(addr_port))
-					{
-						self.p2p_network_service.reserve_peer(*peer_id, *port, *addr);
+					for (peer_id, (port, ip_address)) in peers {
+						self.update_peer_mapping(&mut state, peer_id, port, ip_address);
 					}
 
 					log::info!(
@@ -258,24 +295,19 @@ pub fn new_p2p_validator_network_node<
 				) -> Result<u64> {
 					let peer_id: PeerId = peer_id.try_into()?;
 					let mut state = self.state.write().unwrap();
-					if let Some(port_addr) = state.reserved_peers.get(&peer_id) {
-						if port_addr == &(port, ip_address) {
-							return Err(jsonrpc_core::Error::invalid_params(format!(
-								"Tried to add peer {} which is already reserved",
-								peer_id
-							)))
-						} else {
-							self.p2p_network_service.remove_reserved_peer(peer_id);
-						}
+					if self.update_peer_mapping(&mut state, peer_id, port, ip_address) {
+						log::info!(
+							"Added reserved {} peer (Total Reserved: {})",
+							CHAINFLIP_P2P_PROTOCOL_NAME,
+							state.reserved_peers.len()
+						);
+						Ok(200)
+					} else {
+						Err(jsonrpc_core::Error::invalid_params(format!(
+							"Tried to add peer {} which is already reserved",
+							peer_id
+						)))
 					}
-					state.reserved_peers.insert(peer_id, (port, ip_address));
-					self.p2p_network_service.reserve_peer(peer_id, port, ip_address);
-					log::info!(
-						"Added reserved {} peer (Total Reserved: {})",
-						CHAINFLIP_P2P_PROTOCOL_NAME,
-						state.reserved_peers.len()
-					);
-					Ok(200)
 				}
 
 				/// Disconnect from a validator
@@ -311,8 +343,16 @@ pub fn new_p2p_validator_network_node<
 
 					let state = self.state.read().unwrap();
 					if peers.iter().all(|peer| state.reserved_peers.contains_key(peer)) {
-						for peer in peers {
-							self.p2p_network_service.write_notification(peer, message.clone());
+						for peer_id in peers {
+							let (_, _, message_sender) =
+								state.reserved_peers.get(&peer_id).unwrap();
+							match message_sender.try_send(message.clone()) {
+								Ok(_) => (),
+								Err(TrySendError::Full(_)) => {
+									log::info!("Dropping message for peer {}", peer_id);
+								},
+								Err(_) => panic!(),
+							}
 						}
 						Ok(200)
 					} else {
@@ -378,6 +418,7 @@ pub fn new_p2p_validator_network_node<
 			RpcRequestHandler {
 				state: state.clone(),
 				p2p_network_service: p2p_network_service.clone(),
+				message_sender_spawner,
 				notification_rpc_subscription_manager: SubscriptionManager::new(Arc::new(
 					subscription_task_executor,
 				)),
@@ -503,6 +544,7 @@ mod tests {
 		P2PRpcClient,
 		Arc<RwLock<P2PValidatorNetworkNodeState>>,
 		Arc<LockedMockPeerNetwork>,
+		sc_service::TaskManager,
 	) {
 		let (event_sender, event_receiver) = tokio::sync::mpsc::unbounded_channel();
 
@@ -512,8 +554,15 @@ mod tests {
 			Box::pin(tokio_stream::wrappers::UnboundedReceiverStream::new(event_receiver))
 		});
 
+		let handle = tokio::runtime::Handle::current();
+		let task_executor: sc_service::TaskExecutor =
+			(move |future, _| handle.spawn(future).map(|_| ())).into();
+		let task_manager = sc_service::TaskManager::new(task_executor, None).unwrap();
+		let message_sender_spawn_handle = task_manager.spawn_handle();
+
 		let (rpc_request_handler, p2p_message_handler_future) = new_p2p_validator_network_node(
 			network_expectations.clone(),
+			message_sender_spawn_handle,
 			sc_rpc::testing::TaskExecutor,
 		);
 
@@ -530,7 +579,7 @@ mod tests {
 
 		network_expectations.write().checkpoint();
 
-		(event_sender, client, internal_state, network_expectations)
+		(event_sender, client, internal_state, network_expectations, task_manager)
 	}
 
 	async fn expect_reserve_peer_changes_during_closure<F: Future, C: FnOnce() -> F>(
@@ -582,7 +631,13 @@ mod tests {
 		c().await;
 		network_expectations.write().checkpoint();
 		assert_eq!(
-			internal_state.read().unwrap().reserved_peers,
+			internal_state
+				.read()
+				.unwrap()
+				.reserved_peers
+				.iter()
+				.map(|(peer_id, (port, ip_address, _))| (*peer_id, (*port, *ip_address)))
+				.collect::<BTreeMap<_, _>>(),
 			final_state
 				.into_iter()
 				.map(|(peer_id, port, ip_address)| (
@@ -595,7 +650,7 @@ mod tests {
 
 	#[tokio::test]
 	async fn add_and_remove_peers() {
-		let (_event_sender, client, internal_state, network_expectations) =
+		let (_event_sender, client, internal_state, network_expectations, _task_manager) =
 			new_p2p_validator_network_node_with_test_probes().await;
 
 		let peer_0 = PeerIdTransferable::from(&PeerId::random());
@@ -747,7 +802,7 @@ mod tests {
 
 	#[tokio::test]
 	async fn set_peers() {
-		let (_event_sender, client, internal_state, network_expectations) =
+		let (_event_sender, client, internal_state, network_expectations, _task_manager) =
 			new_p2p_validator_network_node_with_test_probes().await;
 
 		let peer_0 = PeerIdTransferable::from(&PeerId::random());
@@ -843,7 +898,7 @@ mod tests {
 
 	#[tokio::test]
 	async fn send_message() {
-		let (_event_sender, client, _internal_state, network_expectations) =
+		let (_event_sender, client, _internal_state, network_expectations, _task_manager) =
 			new_p2p_validator_network_node_with_test_probes().await;
 
 		let peer_0 = PeerId::random();
@@ -936,7 +991,7 @@ mod tests {
 
 	#[tokio::test]
 	async fn rpc_subscribe() {
-		let (event_sender, client, _internal_state, _expectations) =
+		let (event_sender, client, _internal_state, _expectations, _task_manager) =
 			new_p2p_validator_network_node_with_test_probes().await;
 
 		let peer_0 = PeerId::random();
